requirements.txt
apache-airflow-providers-google==10.9.0
astronomer-cosmos==1.1.3
boto3==1.28.62
pandas==2.1.1